rename : modify SearchSection name to SearchForm
